Use fully qualified name for Integer.
Latest Scala compiler seems to be confused if
we use just `Integer` to refer to java.lang.Integer.

That confusion results in broken output files
(it goes through type checker!) which is bad. Using
fully qualified name is enough to work-aroudn this
issue. Nailing it down would be better but those
things tend to take days to address so I won't
bother at the moment.

Change-Id: Ibd0e6d791277093e5c451a9249bd189818fbf87f
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>